third_party/WebKit/LayoutTests/fast/dom/shadow/style-sharing-host-attribute-expected.txt

CONSOLE WARNING: line 18: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Don't share style between two hosts with different attribute matching for :host([])

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(roots[1]).color is 'rgb(0, 128, 0)'
PASS successfullyParsed is true

TEST COMPLETE

This text should be green
